Big refactorization: 1. Moved methods to service 2. Added handlers 3. Extracted commands 4. Little changes to AuthController
