+++ VeteriesServer.sln
@@ -75,7 +75,7 @@ Global
 		{941D03FC-8CBC-457B-B03B-F9081DA82CA5} = {50AD07A8-8750-4CA5-812B-429A3002F658}
 		{27CE117E-961B-45B8-9D56-3478823B6FD0} = {50AD07A8-8750-4CA5-812B-429A3002F658}
 		{9C47DCBF-4EE0-4B90-AD6A-853945EF658C} = {50AD07A8-8750-4CA5-812B-429A3002F658}
-		{E9522D1D-870E-4F85-9C8F-2649D92351E2} = {C488F066-CEAD-4C24-A98E-DEE6002E540F}
+		{E9522D1D-870E-4F85-9C8F-2649D92351E2} = {50AD07A8-8750-4CA5-812B-429A3002F658}
 		{F95E99FF-5856-4875-AB69-1A4690F7C952} = {27CE117E-961B-45B8-9D56-3478823B6FD0}
 		{86F89E7D-874A-414D-AD42-B54238D49A74} = {941D03FC-8CBC-457B-B03B-F9081DA82CA5}
 		{EFFB397B-C7F7-499F-90AC-AD501322BC32} = {941D03FC-8CBC-457B-B03B-F9081DA82CA5}
