+++ src/Feature/User/DependencyInjection.cs
@@ -4,6 +4,7 @@ using MediatR;
 using System.Reflection;
 using User.Abstractions;
 using User.Abstractions.Internals;
+using User.Internals;
 
 namespace User
 {
@@ -17,6 +18,7 @@ namespace User
             builder.RegisterAssemblyTypes(typeof(DependencyInjection).GetTypeInfo().Assembly)
                 .AsClosedTypesOf(typeof(AbstractValidator<>));
 
+            builder.RegisterType<IdentityService>().As<IIdentityService>();
             builder.RegisterType<TokenService>().As<ITokenService>();
         }
     }
