--- src/Feature/User/Models/AuthenticationResult.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace User.Models
-{
-    public class AuthenticationResult
-    {
-        public string Token { get; set; }
-        public bool Success { get; set; }
-        public IEnumerable<string> ErrorMessages { get; set; }
-    }
-}
