+++ src/Feature/User/Models/Commands/AuthenticateUserCommand.cs
@@ -1,0 +1,11 @@
+﻿using MediatR;
+using User.Models.Results;
+
+namespace User.Models.Commands
+{
+    public class AuthenticateUserCommand : IdentityCommand, IRequest<IdentityResult>
+    {
+        public string Email { get; set; }
+        public string Password { get; set; }
+    }
+}
