+++ src/Feature/User/Models/Commands/IdentityCommand.cs
@@ -1,0 +1,10 @@
+﻿namespace User.Models.Commands
+{
+    public class IdentityCommand
+    {
+        public bool IsNull()
+        {
+            return this == null ? true : false;
+        }
+    }
+}
