+++ src/Feature/User/Models/Commands/RefreshTokenCommand.cs
@@ -1,0 +1,11 @@
+﻿using MediatR;
+using User.Models.Results;
+
+namespace User.Models.Commands
+{
+    public class RefreshTokenCommand : IdentityCommand, IRequest<IdentityResult>
+    {
+        public string Token { get; set; }
+        public string RefreshToken { get; set; }
+    }
+}
